+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implemented download dir
 - Code split the solution
 - Base refactor modules concept
+- Base Health checker
 
 ### Modified